requirements.txt
numpy
scipy
pyaudio
PyFT8
pyserial
